Move player ghost town code to separate component
