+++ src/app/player/player.component.scss
@@ -11,12 +11,6 @@
   pointer-events: none !important;
 }
 
-.ghost-towns-table {
-  .cell {
-    max-width: unset !important;
-  }
-}
-
 mat-expansion-panel-header {
   height: 40px !important;
 
@@ -99,11 +93,4 @@ p {
 }
 mat-chip:hover {
   cursor: pointer !important;
-}
-
-.ghost-won {
-  background: #00961050 !important;
-}
-.ghost-lost {
-  background: #CF443550 !important;
 }
